# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Yamo Laundry API
  version: 1.0.1
  description: >-
    REST API for the Yamo laundry platform. Includes customer, orders,
    tasks, billing, geo, notifications, livreur management and live tracking.
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://api.yamo.example
    description: Production
...
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
              required: [username, password]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user:
                    type: object
                    properties:
                      id: { type: integer, format: int64 }
                      username: { type: string }
                      role: { $ref: '#/components/schemas/Role' }
                      customerId: { type: integer, format: int64, nullable: true }
                      livreurId: { type: integer, format: int64, nullable: true }
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Create user (bootstrap/admin only in production)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                role: { $ref: '#/components/schemas/Role' }
                customerId: { type: integer, format: int64, nullable: true }
                livreurId: { type: integer, format: int64, nullable: true }
              required: [username, password]
      responses:
        '200': { description: Created }
  /api/customers:
    get:
      tags: [Customers]
      summary: List customers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      tags: [Customers]
      summary: Create customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /api/customers/{id}:
    get:
      tags: [Customers]
      summary: Get customer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404': { description: Not found }
    delete:
      tags: [Customers]
      summary: Delete customer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: No content }
  /api/customers/{customerId}/addresses:
    get:
      tags: [Addresses]
      summary: List addresses for a customer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customerId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
    post:
      tags: [Addresses]
      summary: Add address for a customer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customerId
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Address' }
  /api/addresses/{id}:
    delete:
      tags: [Addresses]
      summary: Delete address
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: No content }
...
  /api/tasks/customer/{customerId}:
    get:
      tags: [Tasks]
      summary: List tasks for a customer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customerId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
...
  /api/billing/orders/{orderId}/invoice:
    post:
      tags: [Billing]
      summary: Create invoice for order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: dueDate
          required: true
          schema: { type: string, format: date }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
  /api/billing/customers/{customerId}/invoices:
    get:
      tags: [Billing]
      summary: List invoices for a customer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customerId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Invoice' } } } } }
  /api/billing/customers/{customerId}/payments:
    get:
      tags: [Billing]
      summary: List payments for a customer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customerId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Payment' } } } } }
  /api/billing/payments:
    post:
      tags: [Billing]
      summary: Record payment
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: customerId
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: amount
          required: true
          schema: { type: number, format: double }
        - in: query
          name: method
          required: true
          schema: { $ref: '#/components/schemas/PaymentMethod' }
        - in: query
          name: reference
          required: false
          schema: { type: string }
        - in: query
          name: invoiceNumber
          required: false
          schema: { type: string }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Payment' } } } }
...
  /api/billing/invoices/{invoiceNumber}/pdf:
    post:
      tags: [Billing]
      summary: Generate invoice PDF and return its URL
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: invoiceNumber
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceNumber: { type: string }
                  file: { type: string }
                  url: { type: string, format: uri }
  /api/billing/invoices/pdf/{invoiceNumber}:
    get:
      tags: [Billing]
      summary: Download invoice PDF
      parameters:
        - in: path
          name: invoiceNumber
          required: true
          schema: { type: string }
      responses:
        '200':
          description: PDF content
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /api/billing/invoices/{invoiceNumber}/send-whatsapp:
    post:
      tags: [Billing]
      summary: Send invoice via WhatsApp document link
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: invoiceNumber
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/geo/coverage:
    post:
      tags: [Geo]
      summary: Check coverage and distance to center
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LatLng' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoverageResult' }
  /api/geo/estimate:
    post:
      tags: [Geo]
      summary: Estimate distance, duration and fee between two points
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EstimateReq' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstimateResult' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Role:
      type: string
      enum: [CUSTOMER, LIVREUR, ADMIN]
    Customer:
      type: object
      properties:
        id: { type: integer, format: int64 }
        code: { type: string }
        name: { type: string }
        surname: { type: string }
        phone: { type: string }
        whatsappPhone: { type: string }
        email: { type: string, format: email }
        address1: { type: string }
        address2: { type: string }
        nui: { type: string }
        rccm: { type: string }
        companyName: { type: string }
        segment: { type: string }
        discountPercentage: { type: number, format: double }
        creditBalance: { type: number, format: double }
    Address:
      type: object
      properties:
        id: { type: integer, format: int64 }
        customer: { $ref: '#/components/schemas/CustomerRef' }
        name: { type: string }
        district: { $ref: '#/components/schemas/District' }
        street: { type: string }
        buildingNumber: { type: string }
        apartmentNumber: { type: string }
        additionalInstructions: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        primary: { type: boolean }
    CustomerRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
    District:
      type: string
      description: Named district enumeration
...
    Task:
      type: object
      properties:
        id: { type: integer, format: int64 }
        type: { $ref: '#/components/schemas/TaskType' }
        order: { $ref: '#/components/schemas/OrderRef' }
        customer: { $ref: '#/components/schemas/CustomerRef' }
        title: { type: string }
        description: { type: string }
        addressLine: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        scheduledAt: { type: string, format: date-time }
        remindBeforeMinutes: { type: integer }
        assigneeLivreur: { $ref: '#/components/schemas/LivreurRef' }
        participantsPhones:
          type: array
          items: { type: string }
        status: { $ref: '#/components/schemas/TaskStatus' }
        notes: { type: string }
        proofPhotoUrl: { type: string }
        proofSignatureUrl: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TaskType:
      type: string
      enum: [PICKUP, DELIVERY, OTHER]
    TaskStatus:
      type: string
      enum: [SCHEDULED, IN_PROGRESS, DONE, CANCELLED, MISSED]
    OrderRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
...
    LivreurRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
...
    LatLng:
      type: object
      properties:
        lat: { type: number, format: double }
        lng: { type: number, format: double }
    CoverageResult:
      type: object
      properties:
        serviceable: { type: boolean }
        message: { type: string }
        distanceFromCenterKm: { type: number, format: double }
    EstimateReq:
      type: object
      properties:
        originLat: { type: number, format: double }
        originLng: { type: number, format: double }
        destLat: { type: number, format: double }
        destLng: { type: number, format: double }
    EstimateResult:
      type: object
      properties:
        distanceMeters: { type: integer, format: int64 }
        durationSeconds: { type: integer, format: int64 }
        distanceKm: { type: number, format: double }
        fee: { type: number, format: double }
    Invoice:
      type: object
      properties:
        id: { type: integer, format: int64 }
        invoiceNumber: { type: string }
        issueDate: { type: string, format: date-time }
        dueDate: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/InvoiceStatus' }
        totalAmount: { type: number, format: double }
        paidAmount: { type: number, format: double }
        balanceAmount: { type: number, format: double }
        customer: { $ref: '#/components/schemas/CustomerRef' }
    InvoiceStatus:
      type: string
      enum: [DRAFT, ISSUED, PARTIALLY_PAID, PAID, CANCELLED]
    Payment:
      type: object
      properties:
        id: { type: integer, format: int64 }
        customer: { $ref: '#/components/schemas/CustomerRef' }
        invoiceNumber: { type: string }
        amount: { type: number, format: double }
        method: { $ref: '#/components/schemas/PaymentMethod' }
        reference: { type: string }
        createdAt: { type: string, format: date-time }
    PaymentMethod:
      type: string
      enum: [OM, MOMO, ESPECES, VIREMENT]
...
security:
  - bearerAuth: []